% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_data_sim.R
\docType{data}
\name{all_data_sim}
\alias{all_data_sim}
\title{All Nightly Operations Data Simulated}
\format{a data frame
\describe{
  \item{Day}{Day of week: Factor}
  \item{FY}{Fiscal Year: Factor}
  \item{Week}{Week of Semester: Factor}
  \item{Rides}{Number of rides given that night: Numeric}
  }}
\source{
This data is simulated to represent the data used
}
\usage{
all_data_sim
}
\description{
Simulated data points from first 8 weeks of Fiscal Year 2014 to Fiscal Year 2020. Operations are on the Thursday, Friday, and Saturday of each week. The total number of rides serviced is recorded.
}
\keyword{datasets}
