% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANSIM.R, R/all_data_sim.R
\docType{data}
\name{all_data_sim}
\alias{all_data_sim}
\title{All Nightly Operations Data Simulated}
\format{a data frame
\describe{
  \item{Day}{Day of week: Factor}
  \item{FY}{Fiscal Year: Factor}
  \item{Week}{Week of Semester: Factor}
  \item{Rides}{Number of rides given that night: numeric}
  }}
\source{
This data is simulated to represent the data used

This data is simulated to represent the data used
}
\usage{
all_data_sim

all_data_sim
}
\description{
Simulated data points from first 8 weeks of fy14-fy19

Simulated data points from first 8 weeks of fy14-fy19
}
\keyword{datasets}
