% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_est.R
\name{post_est}
\alias{post_est}
\title{Run the Model}
\usage{
post_est(XMat, Y, mu, tau2, logmu0, S, depend = TRUE)
}
\arguments{
\item{XMat}{Matrix of input data (n-by-m)}

\item{Y}{Matrix of out put data (n-by-1)}

\item{mu}{Vector of mean used in drawing v (n-by-1)}

\item{tau2}{Variance tau^2 used to control variance of the distribution that v is drawn from}

\item{logmu0}{log(mu_0) Log of the mean of Y, used to center the data}

\item{S}{the number of iterations}

\item{depend}{logicial, is there dependence in the covariance sturcture of the prior distribution?}
}
\value{
The estimated beta (n-by-1 matrix) for my specific model
}
\description{
Run the Model
}
